@@ -1,123 +1,118 @@
 package list
 
-import "fmt"
-import "testing"
+import (
+	"fmt"
+	"testing"
+)
 
-func ExampleList1() {
-	l := New()
+func Example_list1() {
+	l := New[string]()
 	l.PushBack("ddddd")
 	l.PushFront("ccccc")
 	l.PushBack("aaaaa")
 	l.PushFront("bbbbb")
-	
-	l.RangeFront(func(v interface{}) bool {
+
+	l.RangeFront(func(v string) bool {
 		fmt.Printf("%v\n", v)
 		return true
 	})
 	fmt.Printf("%v\n", l.Size())
-/* Output:
-bbbbb
-ccccc
-ddddd
-aaaaa
-4
-*/
+	// Output:
+	// bbbbb
+	// ccccc
+	// ddddd
+	// aaaaa
+	// 4
 }
 
-type MyLess_t struct {}
-
-func (MyLess_t) Less(a interface{}, b interface{}) bool {
-	return a.(string) < b.(string)
+func LessStr(a string, b string) bool {
+	return a < b
 }
 
-func ExampleList2() {
-	l := New()
+func Example_list2() {
+	l := New[string]()
 	l.PushBack("ddddd")
 	l.PushFront("ccccc")
 	l.PushBack("aaaaa")
 	l.PushFront("bbbbb")
-	
-	l.InsertionSortFront(MyLess_t{})
-	l.RangeFront(func(v interface{}) bool {
+
+	l.InsertionSortFront(LessStr)
+	l.RangeFront(func(v string) bool {
 		fmt.Printf("%v\n", v)
 		return true
 	})
 	fmt.Printf("%v\n", l.Size())
-/* Output:
-aaaaa
-bbbbb
-ccccc
-ddddd
-4
-*/
+	// Output:
+	// aaaaa
+	// bbbbb
+	// ccccc
+	// ddddd
+	// 4
 }
 
-func ExampleList3() {
-	l := New()
+func Example_list3() {
+	l := New[string]()
 	l.PushBack("ddddd")
 	l.PushFront("ccccc")
 	l.PushBack("aaaaa")
 	l.PushFront("bbbbb")
-	
-	l.InsertionSortBack(MyLess_t{})
-	l.RangeFront(func(v interface{}) bool {
+
+	l.InsertionSortBack(LessStr)
+	l.RangeFront(func(v string) bool {
 		fmt.Printf("%v\n", v)
 		return true
 	})
 	fmt.Printf("%v\n", l.Size())
-/* Output:
-ddddd
-ccccc
-bbbbb
-aaaaa
-4
-*/
+	// Output:
+	// ddddd
+	// ccccc
+	// bbbbb
+	// aaaaa
+	// 4
 }
 
-func ExampleList4() {
-	l := New()
+func Example_list4() {
+	l := New[string]()
 	l.PushBack("ddddd")
 	l.PushFront("ccccc")
 	l.PushBack("aaaaa")
 	l.PushFront("bbbbb")
-	
-	l.InsertionSortFront(MyLess_t{})
-	l.RangeBack(func(v interface{}) bool {
+
+	l.InsertionSortFront(LessStr)
+	l.RangeBack(func(v string) bool {
 		fmt.Printf("%v\n", v)
 		return true
 	})
 	fmt.Printf("%v\n", l.Size())
-/* Output:
-ddddd
-ccccc
-bbbbb
-aaaaa
-4
-*/
+	// Output:
+	// ddddd
+	// ccccc
+	// bbbbb
+	// aaaaa
+	// 4
 }
 
-func ExampleList5() {
-	l := New()
+func Example_list5() {
+	l := New[string]()
 	l.PushBack("ddddd")
 	l.PushFront("ccccc")
 	l.PushBack("aaaaa")
 	l.PushFront("bbbbb")
-	
-	l.InsertionSortBack(MyLess_t{})
-	l.RangeBack(func(v interface{}) bool {
+
+	l.InsertionSortBack(LessStr)
+	l.RangeBack(func(v string) bool {
 		fmt.Printf("%v\n", v)
 		return true
 	})
 	fmt.Printf("%v\n", l.Size())
-/* Output:
-aaaaa
-bbbbb
-ccccc
-ddddd
-4
-*/
+	// Output:
+	// aaaaa
+	// bbbbb
+	// ccccc
+	// ddddd
+	// 4
 }
 
-func Test1(t * testing.T) {
+func Test1(t *testing.T) {
 
 }
